ASCII art verplaatst van Startscherm naar ASCIIART.cs
